docs: Add OpenAI API compatibility page
This adds some initial content documenting our OpenAI compatible APIs
- Responses, Chat Completions, Completions, and Models - along with
instructions on how to use them via OpenAI or Llama Stack clients and
some simple examples for each.

It's not a lot of content, but it's a start so that users have some
idea how to get going as we continue to work on these APIs.

Signed-off-by: Ben Browning <bbrownin@redhat.com>