chore: cleanups from review feedback on openai api docs
This moves the Models section to the top of the supported APIs, along
with a brief explanation so that it's not just a bare section with
just a small code sample.

This also adjusts the MyST directives to use code fences instead of
markdown directives, as that allows word wrapping to work as normal
instead of it rendering as a single long line in the GitHub UI and
such. The actual rendered HTML content is identical, but this makes it
a bit easier to review.

And, the warning from the Responses API structured output section is
removed because we've now merged support for that.

Signed-off-by: Ben Browning <bbrownin@redhat.com>