Use VectorStoreContent vs InterleavedContent in vector store files
This extracts the existing logic to convert chunks to
VectorStoreContent objects into a reusable method and uses that when
returning our list of Vector Store File contents.

It also adds an xfail test for deleting vector store files, as that's
not implemented yet but parking the implementation of that for now.

Signed-off-by: Ben Browning <bbrownin@redhat.com>