Add a helpful error message when Fails fails to load indexes
We've had a few cases recently where Faiss is unable to load its
indexes properly, like after users upgrade versions of Llama Stack or
numpy dependencies. This catches any exceptions when deserializing the
index with an error message about what likely went wrong and how to
potentially fix it so that users can attempt to fix this without
needing to raise an issue or being confused.

The most common case of new users hitting this will be with the
SqliteKVStoreImpl, since that's our local storage most people use when
testing out Llama Stack. So, I added a `__str__` method to it that
shows the proper path and table name that users need to potentially
delete to resolve this error.

Signed-off-by: Ben Browning <bbrownin@redhat.com>