fix: Restore the nvidia distro
The `nvidia` distro was previously collapsed into the `starter`
distro. However, the `nvidia` distro was setup specifically to use
NVIDIA NeMo microservices as providers for all APIs and not just
inference, which means it was doing quite a bit more than what the
`starter` distro covers today.

We should work with our friends at NVIDIA to determine the best place
to maintain this distro long-term, but for now this restores the
`nvidia` distro and its docs back to where they were so that things
continue to work for their users.

Signed-off-by: Ben Browning <bbrownin@redhat.com>