feat: llama stack sync
`llama stack sync` uses `uv` so install all required pip dependencies. It does this by using the `module` in the provider spec as the installation directory and runs

`uv pip install $DIR`. Since each provider now has a `pyproject.toml`, they can be installed as a package using the dependencies in the pyproject

Signed-off-by: Charlie Doern <cdoern@redhat.com>