Use `pool_pre_ping=True` in SQLAlchemy engine creation
We noticed that when llama-stack is running for a long time,  we would
run into database errors when trying to run messages through the agent
(which we configured to persist against postgres), seemingly due to the
database connections being stale or disconnected. This commit adds
`pool_pre_ping=True` to the SQLAlchemy engine creation to help mitigate
this issue by checking the connection before using it, and
re-establishing it if necessary.

More information in:

https://docs.sqlalchemy.org/en/20/core/pooling.html#dealing-with-disconnects

We're also open to other suggestions on how to handle this issue, this
PR is just a suggestion, we have not tested it yet (we're in the process
of doing that) and we're hoping it's going to resolve our issue.