test: introduce api conformance test
this test runs on each PR and uses a new conformance workflow to compare the base (main) branch openapi spec to the one on this PR
if one of our "stable" APIs change, the test will fail.

this workflow needs a different version of `yq` than the one that comes with `ubuntu`. Additonally yq has issues writing files to disk in the workflows environment, so the yq itself needs to happen in the same container as the openapi-diff.

This means I needed to write a new Contaninerfile and build a new Container which has two stages: first install yq (since the openapi diff image is distroless) and second use yq to parse the two openapi specs and THEN use openapi-diff to actually diff the two specs.

As a follow up to this, we can add an optional way to make it so that it is OK to make these change if properly documented or in a changelog or something.

related to #3237

Signed-off-by: Charlie Doern <cdoern@redhat.com>